Add files via upload
This is a JavaScript Book Reading Log. It counts the books I am reading, displays my progress by page counts, performs the calculations of my overall reading, and provides the status of each book I am reading. It also presents the most-pages-remaining of the book, and overview of my month-long reading.